mv oldpvtdata commit to separate file
