+++ core/ledger/pvtdatastorage/store.go
@@ -107,12 +107,6 @@ type storeEntries struct {
 // and is stored as the value of lastUpdatedOldBlocksKey (defined in kv_encoding.go)
 type lastUpdatedOldBlocksList []uint64
 
-type entriesForPvtDataOfOldBlocks struct {
-	dataEntries        map[dataKey]*rwset.CollectionPvtReadWriteSet
-	expiryEntries      map[expiryKey]*ExpiryData
-	missingDataEntries map[nsCollBlk]*bitset.BitSet
-}
-
 //////// Provider functions  /////////////
 //////////////////////////////////////////
 
@@ -260,206 +254,6 @@ func (s *Store) Commit(blockNum uint64, pvtData []*ledger.TxPvtData, missingPvtD
 	atomic.StoreUint64(&s.lastCommittedBlock, committingBlockNum)
 	logger.Debugf("Committed private data for block [%d]", committingBlockNum)
 	s.performPurgeIfScheduled(committingBlockNum)
-	return nil
-}
-
-// CommitPvtDataOfOldBlocks commits the pvtData (i.e., previously missing data) of old blocks.
-// The parameter `blocksPvtData` refers a list of old block's pvtdata which are missing in the pvtstore.
-// Given a list of old block's pvtData, `CommitPvtDataOfOldBlocks` performs the following three
-// operations
-// (1) construct update entries (i.e., dataEntries, expiryEntries, missingDataEntries)
-//     from the above created data entries
-// (2) create a db update batch from the update entries
-// (3) commit the update batch to the pvtStore
-func (s *Store) CommitPvtDataOfOldBlocks(blocksPvtData map[uint64][]*ledger.TxPvtData) error {
-	if s.isLastUpdatedOldBlocksSet {
-		return &ErrIllegalCall{`The lastUpdatedOldBlocksList is set. It means that the
-		stateDB may not be in sync with the pvtStore`}
-	}
-
-	logger.Debugf("Constructing pvtdatastore entries for pvtData of [%d] old blocks", len(blocksPvtData))
-	entries, err := s.constructEntries(blocksPvtData)
-	if err != nil {
-		return err
-	}
-
-	logger.Debug("Constructing update batch from pvtdatastore entries")
-	batch := s.db.NewUpdateBatch()
-	if err := entries.addToUpdateBatch(batch); err != nil {
-		return err
-	}
-
-	logger.Debug("Committing the update batch to pvtdatastore")
-	return s.db.WriteBatch(batch, true)
-}
-
-func (s *Store) constructEntries(blocksPvtData map[uint64][]*ledger.TxPvtData) (*entriesForPvtDataOfOldBlocks, error) {
-	var dataEntries []*dataEntry
-	for blkNum, pvtData := range blocksPvtData {
-		dataEntries = append(dataEntries, prepareDataEntries(blkNum, pvtData)...)
-	}
-
-	entries := &entriesForPvtDataOfOldBlocks{
-		dataEntries:        make(map[dataKey]*rwset.CollectionPvtReadWriteSet),
-		expiryEntries:      make(map[expiryKey]*ExpiryData),
-		missingDataEntries: make(map[nsCollBlk]*bitset.BitSet),
-	}
-
-	for _, dataEntry := range dataEntries {
-		var expData *ExpiryData
-		nsCollBlk := dataEntry.key.nsCollBlk
-		txNum := dataEntry.key.txNum
-
-		expKey, err := s.constructExpiryKeyFromDataEntry(dataEntry)
-		if err != nil {
-			return nil, err
-		}
-		if !neverExpires(expKey.expiringBlk) {
-			if expData, err = s.getExpiryDataFromEntriesOrStore(entries, expKey); err != nil {
-				return nil, err
-			}
-			if expData == nil {
-				// data entry is already expired
-				// and purged (a rare scenario)
-				continue
-			}
-			expData.addPresentData(nsCollBlk.ns, nsCollBlk.coll, txNum)
-		}
-
-		var missingData *bitset.BitSet
-		if missingData, err = s.getMissingDataFromEntriesOrStore(entries, nsCollBlk); err != nil {
-			return nil, err
-		}
-		if missingData == nil {
-			// data entry is already expired
-			// and purged (a rare scenario)
-			continue
-		}
-		missingData.Clear(uint(txNum))
-
-		entries.add(dataEntry, expKey, expData, missingData)
-	}
-	return entries, nil
-}
-
-func (s *Store) constructExpiryKeyFromDataEntry(dataEntry *dataEntry) (expiryKey, error) {
-	// get the expiryBlk number to construct the expiryKey
-	nsCollBlk := dataEntry.key.nsCollBlk
-	expiringBlk, err := s.btlPolicy.GetExpiringBlock(nsCollBlk.ns, nsCollBlk.coll, nsCollBlk.blkNum)
-	if err != nil {
-		return expiryKey{}, err
-	}
-
-	return expiryKey{
-		expiringBlk:   expiringBlk,
-		committingBlk: nsCollBlk.blkNum,
-	}, nil
-}
-
-func (s *Store) getExpiryDataFromEntriesOrStore(updateEntries *entriesForPvtDataOfOldBlocks, expiryKey expiryKey) (*ExpiryData, error) {
-	if expiryData, ok := updateEntries.expiryEntries[expiryKey]; ok {
-		return expiryData, nil
-	}
-
-	expiryData, err := s.getExpiryDataOfExpiryKey(&expiryKey)
-	if err != nil {
-		return nil, err
-	}
-	return expiryData, nil
-}
-
-func (s *Store) getMissingDataFromEntriesOrStore(updateEntries *entriesForPvtDataOfOldBlocks, nsCollBlk nsCollBlk) (*bitset.BitSet, error) {
-	if missingData, ok := updateEntries.missingDataEntries[nsCollBlk]; ok {
-		return missingData, nil
-	}
-
-	missingDataKey := &missingDataKey{
-		nsCollBlk:  nsCollBlk,
-		isEligible: true,
-	}
-	missingData, err := s.getBitmapOfMissingDataKey(missingDataKey)
-	if err != nil {
-		return nil, err
-	}
-	return missingData, nil
-}
-
-func (e *entriesForPvtDataOfOldBlocks) add(datEntry *dataEntry, expKey expiryKey, expData *ExpiryData, missingData *bitset.BitSet) {
-	dataKey := dataKey{
-		nsCollBlk: datEntry.key.nsCollBlk,
-		txNum:     datEntry.key.txNum,
-	}
-	e.dataEntries[dataKey] = datEntry.value
-
-	if expData != nil {
-		e.expiryEntries[expKey] = expData
-	}
-
-	e.missingDataEntries[dataKey.nsCollBlk] = missingData
-}
-
-func (e *entriesForPvtDataOfOldBlocks) addToUpdateBatch(batch *leveldbhelper.UpdateBatch) error {
-	if err := e.addDataEntriesToUpdateBatch(batch); err != nil {
-		return err
-	}
-
-	if err := e.addExpiryEntriesToUpdateBatch(batch); err != nil {
-		return err
-	}
-
-	return e.addMissingDataEntriesToUpdateBatch(batch)
-}
-
-func (e *entriesForPvtDataOfOldBlocks) addDataEntriesToUpdateBatch(batch *leveldbhelper.UpdateBatch) error {
-	var key, val []byte
-	var err error
-
-	for dataKey, pvtData := range e.dataEntries {
-		key = encodeDataKey(&dataKey)
-		if val, err = encodeDataValue(pvtData); err != nil {
-			return err
-		}
-		batch.Put(key, val)
-	}
-	return nil
-}
-
-func (e *entriesForPvtDataOfOldBlocks) addExpiryEntriesToUpdateBatch(batch *leveldbhelper.UpdateBatch) error {
-	var key, val []byte
-	var err error
-
-	for expiryKey, expiryData := range e.expiryEntries {
-		key = encodeExpiryKey(&expiryKey)
-		if val, err = encodeExpiryValue(expiryData); err != nil {
-			return err
-		}
-		batch.Put(key, val)
-	}
-	return nil
-}
-
-func (e *entriesForPvtDataOfOldBlocks) addMissingDataEntriesToUpdateBatch(batch *leveldbhelper.UpdateBatch) error {
-	var key, val []byte
-	var err error
-
-	for nsCollBlk, missingData := range e.missingDataEntries {
-		key = encodeMissingDataKey(
-			&missingDataKey{
-				nsCollBlk:  nsCollBlk,
-				isEligible: true,
-			},
-		)
-
-		if missingData.None() {
-			batch.Delete(key)
-			continue
-		}
-
-		if val, err = encodeMissingDataValue(missingData); err != nil {
-			return err
-		}
-		batch.Put(key, val)
-	}
 	return nil
 }
